+++ tests/conftest.py
@@ -29,6 +29,7 @@ pytest_plugins = [
     "tests.api.v2_3_7_6_1",
     "tests.api.v2_3_7_9",
     "tests.api.v3_1_3_0",
+    "tests.api.v3_1_6_0",
 ]